requirements.txt
aldryn-apphooks-config==0.6.0
aldryn-boilerplates==0.8.0
aldryn-forms==6.2.1
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.0
cssselect==1.1.0
cssutils==2.3.0
Django==3.2.4
django-absolute==0.3
django-appconf==1.0.4
django-appdata==0.3.2
django-classy-tags==2.0.0
django-cms==3.9.0
django-ddm==0.1.16
django-emailit==0.2.4
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-meta==2.0.0
django-mptt==0.12.0
django-parler==2.2
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-sizefield==1.0.0
django-sortedm2m==3.0.2
django-tablib==3.2
django-taggit==1.5.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.0.0
djangocms-blog==1.2.3
djangocms-text-ckeditor==4.0.0
django-webpack-loader==1.4.1
easy-thumbnails==2.7.1
html5lib==1.1
idna==3.2
importlib-metadata==4.6.1
lxml==4.6.3
mysqlclient==2.1.0
Pillow==8.2.0
premailer==3.9.0
psycopg2-binary==2.9.1
pycryptodome==3.14.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.10.0.0
Unidecode==1.1.2
urllib3==1.26.6
webencodings==0.5.1
YURL==1.0.0
zipp==3.5.0
